Use abstract `Driver` (#3281)
