+++ src/py/flwr/server/compat/driver_client_proxy.py
@@ -25,7 +25,7 @@ from flwr.common import serde
 from flwr.proto import driver_pb2, node_pb2, task_pb2  # pylint: disable=E0611
 from flwr.server.client_proxy import ClientProxy
 
-from ..driver.driver import GrpcDriverHelper
+from ..driver.grpc_driver import GrpcDriverHelper
 
 SLEEP_TIME = 1
